requirements.txt
audioop-lts; python_version >= "3.13"
asyncpg
discord
numpy
psycopg
python-dotenv
sqlalchemy